Refactors portfolio blog files locations and path references
