+++ src/app/top/portfolio/Portfolio.jsx
@@ -10,9 +10,9 @@ import { ContentItemsCount } from './ContentItemsCount'
 import { Featured } from './featured/Featured'
 import { Blog } from './blog/Blog'
 import { Applications } from './applications/Applications'
-import featuredContentData from '../../../global/data/featured-content.json'
-import blogPostsData from '../../../global/data/blogPosts/blog-posts-list'
-import applicationsData from '../../../global/data/applications'
+import featuredContentData from './featured/data/featured-content'
+import blogPostsData from './blog/data/blog-posts-list'
+import applicationsData from './applications/data/applications'
 import './style/portfolio'
 
 const routerParams = {
