+++ src/app/top/portfolio/blog/BlogPostPreviews.jsx
@@ -4,7 +4,7 @@ import { BlogPostPreview } from './BlogPostPreview'
 export const BlogPostPreviews = ({ blogPostsList, blogPostsDataPath }) => {
   const blogPostPreviews = blogPostsList.map(post => {
     let blogPostDataPath = blogPostsDataPath + post.path
-    let blogPostData = require('../../../' + blogPostDataPath + '/' + post.id)
+    let blogPostData = require('./' + blogPostDataPath + '/' + post.id)
     return (
       <div key={post.id}>
         <BlogPostPreview
