+++ src/app/top/portfolio/blog/data/blogPosts/20200226webpack/webpack-static-assets.json
