--- src/global/data/external-links.json
@@ -1,47 +1,0 @@
-{
-  "Axios": "https://github.com/axios/axios#axios",
-  "Body Parser": "https://www.npmjs.com/package/body-parser",
-  "Cartify": "https://github.com/GalenkoEugene/cartify#cartify",
-  "Capacitor": "https://capacitor.ionicframework.com/",
-  "Coveralls": "https://coveralls.io/",
-  "Cucumber": "https://cucumber.io",
-  "Cypress": "https://www.cypress.io/features",
-  "Devise": "https://github.com/plataformatec/devise",
-  "Devise Token Auth": "https://github.com/lynndylanhurley/devise_token_auth#devise-token-auth",
-  "Enzyme": "https://airbnb.io/enzyme/",
-  "Express": "https://expressjs.com/",
-  "Express-Handlebars": "https://www.npmjs.com/package/express-handlebars",
-  "Formspree": "https://formspree.io/",
-  "fs": "https://nodejs.org/api/fs.html",
-  "Git": "https://git-scm.com/",
-  "Github": "https://github.com/features",
-  "Github-Pages": "https://help.github.com/en/categories/github-pages-basics",
-  "Github-Netlify Checks": "https://docs.netlify.com/site-deploys/notifications/#github-commit-checks",
-  "Haml": "http://haml.info/",
-  "Heroku": "https://www.heroku.com/platform",
-  "Heroku Buildpacks": "https://elements.heroku.com/buildpacks",
-  "http": "https://nodejs.org/api/http.html",
-  "Ionic": "https://ionicframework.com/docs",
-  "Jira": "https://www.atlassian.com/software/jira",
-  "jQuery": "https://jquery.com/",
-  "JSON": "https://en.wikipedia.org/wiki/JSON",
-  "Material-UI": "https://material-ui.com/",
-  "Mocha": "https://mochajs.org/",
-  "Netlify": "https://www.netlify.com/products/",
-  "Node": "https://nodejs.org/en/about/",
-  "NPM": "https://www.npmjs.com/",
-  "Pivotal Tracker": "https://www.pivotaltracker.com/",
-  "PostgreSQL": "https://www.postgresql.org/",
-  "Rails": "https://rubyonrails.org/",
-  "Rails Active Storage": "https://edgeguides.rubyonrails.org/active_storage_overview.html",
-  "React": "https://reactjs.org/",
-  "React Router": "https://reacttraining.com/react-router/web/guides/quick-start",
-  "Redux": "https://redux.js.org/",
-  "Ruby_webshot": "https://rubygems.org/gems/ruby_webshot",
-  "Rspec": "https://rspec.info/",
-  "Semantic-UI": "https://semantic-ui.com/usage/layout.html",
-  "Semaphore CI": "https://semaphoreci.com/product",
-  "sessionStorage": "https://developer.mozilla.org/en-US/docs/Web/API/Window/sessionStorage",
-  "Stripe": "https://stripe.com/docs/api",
-  "Xcode": "https://developer.apple.com/xcode/"
-}
