+++ src/index.js
@@ -4,7 +4,7 @@ import { BrowserRouter } from "react-router-dom"
 import { Top } from './app/top/Top'
 import "./app/global/icons/style/icons"
 import "./app/global/links/style/links"
-import './global/fonts/arial'
+import './app/top/style/fonts/arial'
 
 ReactDOM.render(<BrowserRouter> <Top /> </BrowserRouter>,
   document.getElementById("app"));
